+++ dhunter/hunter/args.py
@@ -20,15 +20,29 @@ from ..core.const import Const
 class Args(ArgsBase):
     """Handles command line arguments"""
 
+    def _get_tool_name(self) -> str:
+        return 'dhunt'
+
     # noinspection PyTypeChecker
     def parse_args(self) -> argparse.Namespace:
         """Parses command line arguments
         """
         parser = self._get_parser()
 
-        group = parser.add_argument_group('Project')
-        group.add_argument(metavar='FILE', action='store', dest='db_file', nargs=1,
-                           help='Name of project file to process.')
+        group = parser.add_argument_group('Commands')
+        group.add_argument(metavar='COMMAND', action='store', dest='command', nargs=1,
+                           help='Operation modes: ' + ', '.join(Const.HUNDER_CMDS))
+        # group.add_argument('-c', '-cmd', '--cmd',
+        #                    metavar='MODE', action='store', dest='command', nargs=1,
+        #                    help='Operation modes: ' + ', '.join(Const.HUNDER_CMDS))
+
+        # group.add_argument('-clean-db', '--clean-db',
+        #                    action='store_true', dest='clean_db',
+        #                    help='Updates project db and purges all records for files that no longer exist.')
+
+        group = parser.add_argument_group('Project database')
+        group.add_argument(metavar='DB_FILE', action='store', dest='db_file', nargs=1,
+                           help='Name of project database file to use.')
 
         group = parser.add_argument_group('Sorting and results limit')
         group.add_argument('-s', '-sort', '--sort',