+++ dhunter/scanner/args.py
@@ -20,6 +20,9 @@ from ..core.const import Const
 class Args(ArgsBase):
     """Handles command line arguments"""
 
+    def _get_tool_name(self) -> str:
+        return 'dscan'
+
     # noinspection PyTypeChecker
     def parse_args(self) -> argparse.Namespace:
         """Parses command line arguments
@@ -28,23 +31,32 @@ class Args(ArgsBase):
 
         group = parser.add_argument_group('Project')
         group.add_argument('-db', '--db',
-                           metavar='FILE', action='store', dest='db_file', nargs=1,
-                           help='Name of project file to create.')
+                           metavar='DB_FILE', action='store', dest='db_file', nargs=1,
+                           help='Name of project database file to create.')
 
         group.add_argument(
                 metavar='DIR', action='store', dest='src_dirs', nargs='+',
-                help='Directories to scan and include in the project file.')
+                help='Directories to process.')
 
         group = parser.add_argument_group('Misc')
         group.add_argument('-r', '-rehash', '--rehash',
                            action='store_true', dest='force_rehash',
                            help='Ignores existing file hash cache files ({dot}) and regenerates '
                                 'all the hashes.'.format(dot=Const.FILE_DOT_DHUNTER))
+        group.add_argument('-cmd', '--cmd', '--command',
+                           action='store', dest='command', nargs=1, metavar="COMMAND", default=Const.CMD_SCAN,
+                           help='Action to execute. Commands are: %s. ' % ', '.join(Const.SCANNER_CMDS) +
+                           'Default command is %s' % Const.CMD_SCAN)
         group.add_argument('-ro', '--ro', '--read-only',
                            action='store_true', dest='dont_save_dot_file',
                            help='Tells the scanner to treat all the folders as read only and do not '
-                                'write cache file ({dot})'.format(dot=Const.FILE_DOT_DHUNTER)
-                           )
+                                'write cache file ({dot})'.format(dot=Const.FILE_DOT_DHUNTER))
+        group.add_argument('-rp', '--rp', '--relative-paths',
+                           action='store_true', dest='relative_paths',
+                           help='Store paths in DB as relative instead of absolute.')
+        group.add_argument('-nr', '--nr', '--no-recursive',
+                           action='store_true', dest='no_recursive',
+                           help='Do not scan subdirectories.')
 
         self._add_filter_option_group(parser)
         self._add_other_option_group(parser)