+++ docs/CHANGES.md
@@ -3,6 +3,16 @@
 
 ## Changelog ##
 
+dev
+---
+ * Now uses full absolute path for each item strored in project db file
+ * dhunter: automatically removes dead entries from used DB if they hit the filter
+ * dhunter: added `cleandb` command that removes dead entries from project DB
+ * dhunter: now yields proper warning when source dir is a symlink
+ * dscan: now supports `--force` to overwrite existing database file
+ * dscan: added `--relative-paths` option
+ * dscan: added `--no-recursive` option
+
 v1.2.0 (2019-03-11)
 -------------------
  * Added project logo